Rename hydration reminder worker
